chore: refactor authentication flow
