+++ .eslintrc.json
@@ -2,7 +2,8 @@
   "root": true,
   "plugins": [
     "@typescript-eslint",
-    "filenames"
+    "filenames",
+    "react-hooks"
   ],
   "extends": [
     "next/core-web-vitals",
@@ -40,8 +41,8 @@
     "@typescript-eslint/no-empty-interface": "error",
     "@typescript-eslint/no-empty-function": "error",
     "@typescript-eslint/no-misused-new": "error",
-    "@typescript-eslint/no-var-requires": "error"
-
-    //    "filenames/match-regex": [2, "^[a-z]+(-[a-z]+)*(.d)?$", true]
+    "@typescript-eslint/no-var-requires": "error",
+    "react-hooks/rules-of-hooks": "error",
+    "react-hooks/exhaustive-deps": "warn"
   }
 }
