+++ app/(auth)/register/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from "next"
 
-import DefaultRegisterPage from "@/modules/login/presentation/pages/default/DefaultRegisterPage"
+import DefaultRegisterPage from "@/modules/auth/presentation/pages/default/DefaultRegisterPage"
 
 export const metadata: Metadata = {
   title: "Create an account",
