+++ app/(protected)/(vendor)/logout/page.tsx
@@ -1,0 +1,7 @@
+import React, { ReactElement } from "react"
+
+import DefaultLogoutPage from "@/modules/auth/presentation/pages/default/DefaultLogoutPage"
+
+export default function Page(): ReactElement {
+  return <DefaultLogoutPage />
+}
