+++ middleware.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server"
 
-import { actionGetAuthToken } from "@/modules/login/domain/auth-actions"
+import { actionGetAuthToken } from "@/modules/auth/domain/cookie-service"
 
 export async function middleware(request: NextRequest): Promise<NextResponse> {
   const { nextUrl } = request
@@ -18,14 +18,15 @@ export async function middleware(request: NextRequest): Promise<NextResponse> {
     return NextResponse.redirect(loginUrl)
   }
 
-  if (isAuthentic && isRegisterPage && isLoginPage) {
-    const requestHeaders = new Headers(request.headers)
-    requestHeaders.set("Authorization", `Bearer ${serverToken?.value}`)
+  if (isAuthentic && (isRegisterPage || isLoginPage)) {
+    request.headers.set("Authorization", "Bearer ".concat(serverToken ?? ""))
+
     NextResponse.next({
       request: {
-        headers: requestHeaders,
+        headers: request.headers,
       },
     })
+
     return NextResponse.redirect(new URL("/profile", request.url))
   }
   return NextResponse.next()
