+++ package.json
@@ -83,6 +83,7 @@
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-filenames": "^1.3.2",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "husky": "^8.0.3",
     "prettier": "^3.2.5"
   },
