+++ src/modules/auth/domain/types/auth-config.d.ts
