+++ src/modules/auth/domain/types/index.d.ts
