+++ src/modules/auth/domain/types/nextauth.d.ts
