+++ src/modules/auth/presentation/components/default/client/DefaultLoginProviders.tsx
@@ -1,0 +1,20 @@
+"use client"
+
+import React, { ReactElement } from "react"
+
+import { GithubIcon } from "lucide-react"
+
+import { useAuth } from "@/modules/auth/presentation/hooks/use-auth"
+
+import { Button } from "@/components/ui/button"
+
+export default function DefaultLoginProviders(): ReactElement {
+  const { getLoginProviderLink } = useAuth()
+
+  return (
+    <Button onClick={getLoginProviderLink}>
+      <GithubIcon size="24" />
+      <p className="px-3">Login with Github</p>
+    </Button>
+  )
+}
