+++ src/modules/auth/presentation/components/default/client/DefaultUserRegisterForm.tsx
@@ -1,0 +1,107 @@
+"use client"
+
+import React, { ReactElement } from "react"
+
+import { zodResolver } from "@hookform/resolvers/zod"
+import { Loader2 } from "lucide-react"
+import { useForm } from "react-hook-form"
+
+import {
+  registerFormSchema,
+  RegisterFormValues,
+} from "@/modules/auth/config/form-config"
+import { useAuth } from "@/modules/auth/presentation/hooks/use-auth"
+
+import { Button } from "@/components/ui/button"
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormMessage,
+} from "@/components/ui/form"
+import { Input } from "@/components/ui/input"
+
+export default function DefaultUserRegisterForm(): ReactElement {
+  const { isLoading, registerUser } = useAuth()
+
+  const form = useForm<RegisterFormValues>({
+    resolver: zodResolver(registerFormSchema),
+  })
+
+  return (
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(registerUser)} className="space-y-2">
+        <FormField
+          control={form.control}
+          name="username"
+          render={({ field }): ReactElement => (
+            <FormItem>
+              <FormControl>
+                <Input placeholder="Username" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="password"
+          render={({ field }): ReactElement => (
+            <FormItem>
+              <FormControl>
+                <Input placeholder="Password" type="password" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="confirmPassword"
+          render={({ field }): ReactElement => (
+            <FormItem>
+              <FormControl>
+                <Input placeholder="Password" type="password" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="email"
+          render={({ field }): ReactElement => (
+            <FormItem>
+              <FormControl>
+                <Input placeholder="E-mail" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="phone"
+          render={({ field }): ReactElement => (
+            <FormItem>
+              <FormControl>
+                <Input placeholder="Phone" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <Button disabled={isLoading} className="w-full">
+          {isLoading && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
+          Register
+        </Button>
+      </form>
+    </Form>
+  )
+}
