+++ src/modules/auth/presentation/hooks/index.d.ts
@@ -1,0 +1,9 @@
+import { LoginFormValues } from "@/modules/auth/config/form-config"
+
+export interface IDefaultLoginHooks {
+  isLoading: boolean
+  loginUser: (values: LoginFormValues) => void
+  logoutUser: (values: { revokeAll: boolean }) => void
+
+  getLoginProviderLink: () => void
+}
