+++ src/modules/auth/presentation/models/default/index.d.ts
